first all in one dbutil-plus version

+++ core/src/test/java/com/github/drinkjava2/jtransactions/tinytx/TinyTxTester.java
@@ -46,7 +46,7 @@ public class TinyTxTester {
 	@TX
 	public void tx_Insert2() {
 		tiny.executeSql("insert into users (id) values('456')");
-		Assert.assertEquals(2L, tiny.queryForObject("select count(*) from users"));
+		Assert.assertEquals(2L, (long)tiny.queryForObject("select count(*) from users"));
 		Systemout.println("Now have 2 records in users table, but will roll back to 1");
 		Systemout.println(1 / 0);
 	}
@@ -62,7 +62,7 @@ public class TinyTxTester {
 		}
 		tiny.executeSql("create table users (id varchar(40))engine=InnoDB");
 
-		Assert.assertEquals(0L, tiny.queryForObject("select count(*) from users"));
+		Assert.assertEquals(0L, (long)tiny.queryForObject("select count(*) from users"));
 
 		try {
 			tester.tx_Insert1();// this one inserted 1 record
@@ -70,7 +70,7 @@ public class TinyTxTester {
 		} catch (Exception e) {
 			Systemout.println("div/0 exception should happen, tx_Insert2 should roll back.\r" + e.getMessage());
 		}
-		Assert.assertEquals(1L, tiny.queryForObject("select count(*) from users"));
+		Assert.assertEquals(1L, (long)tiny.queryForObject("select count(*) from users"));
 		JBEANBOX.bctx().close();// Release DataSource Pool
 	}
